move DefaultService table from 'tasks' to 'work_units.zgw' app
