+++ src/bptl/activiti/tests/test_full_process.py
@@ -6,8 +6,8 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from bptl.tasks.models import TaskMapping
-from bptl.tasks.tests.factories import DefaultServiceFactory
 from bptl.utils.constants import Statuses
+from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
 from bptl.work_units.zgw.tests.utils import mock_service_oas_get
 
 from ..models import ServiceTask
