+++ src/bptl/work_units/zgw/tests/test_task_close_zaak.py
@@ -3,7 +3,8 @@ from django.test import TestCase
 import requests_mock
 
 from bptl.camunda.models import ExternalTask
-from bptl.tasks.tests.factories import DefaultServiceFactory, TaskMappingFactory
+from bptl.tasks.tests.factories import TaskMappingFactory
+from bptl.work_units.zgw.tests.factories import DefaultServiceFactory
 
 from ..tasks import CloseZaakTask
 from .utils import mock_service_oas_get
